requirements.txt
Python==3.10.8
pygame==2.1.2
PyOpenGL==3.1.6
numpy==1.23.4
SDL==2.0.18